REST Time & Auth tests
