+++ spec/support/browser_file_list.js
@@ -1,2 +1,2 @@
 window.__karma__ = { base: '../' };
-window.__karma__.files = {"browser/static/ably.js":true,"browser/static/ably.min.js":true,"browser/static/ably.noencryption.js":true,"browser/static/ably.noencryption.min.js":true,"browser/static/compat-pubnub.js":true,"browser/static/compat-pubnub.min.js":true,"browser/static/compat-pusher.js":true,"browser/static/compat-pusher.min.js":true,"browser/static/iframe.js":true,"browser/static/iframe.min.js":true,"browser/static/iframe.html":true,"browser/lib/util/base64.js":true,"spec/common/globals/environment.js":true,"spec/common/globals/named_dependencies.js":true,"spec/common/modules/client_module.js":true,"spec/common/modules/shared_helper.js":true,"spec/common/modules/testapp_manager.js":true,"spec/common/modules/testapp_module.js":true,"spec/support/browser_file_list.js":true,"spec/support/browser_setup.js":true,"spec/support/environment.vars.js":true,"spec/support/modules_helper.js":true,"spec/support/tear_down.js":true,"spec/support/test_helper.js":true,"spec/realtime/auth.test.js":true,"spec/rest/auth.test.js":true};
+window.__karma__.files = {"browser/static/ably.js":true,"browser/static/ably.min.js":true,"browser/static/ably.noencryption.js":true,"browser/static/ably.noencryption.min.js":true,"browser/static/compat-pubnub.js":true,"browser/static/compat-pubnub.min.js":true,"browser/static/compat-pusher.js":true,"browser/static/compat-pusher.min.js":true,"browser/static/iframe.js":true,"browser/static/iframe.min.js":true,"browser/static/iframe.html":true,"browser/lib/util/base64.js":true,"spec/common/globals/environment.js":true,"spec/common/globals/named_dependencies.js":true,"spec/common/modules/client_module.js":true,"spec/common/modules/shared_helper.js":true,"spec/common/modules/testapp_manager.js":true,"spec/common/modules/testapp_module.js":true,"spec/support/browser_file_list.js":true,"spec/support/browser_setup.js":true,"spec/support/environment.vars.js":true,"spec/support/modules_helper.js":true,"spec/support/tear_down.js":true,"spec/support/test_helper.js":true,"spec/realtime/auth.test.js":true,"spec/rest/auth.test.js":true,"spec/rest/capability.test.js":true,"spec/rest/time.test.js":true};
