--- test/rest/time.js
@@ -1,4 +1,0 @@
-"use strict";
-var base = require('./common/common').setup();
-var mod = require('./common/time');
-base.addCommonModule(exports, mod);
